+++ src/common/api/mutations/unArchiveItem.js
@@ -1,5 +1,6 @@
 import { gql } from 'graphql-request'
 import { requestGQL } from 'common/utilities/request/request'
+import { FRAGMENT_ITEM } from 'common/api/fragments/fragment.item'
 
 const itemUnArchiveQuery = gql`
   mutation ItemUnArchive($itemId: ID!) {
@@ -16,10 +17,17 @@ const itemUnArchiveQuery = gql`
         id
         name
       }
+      item {
+        ...ItemDetails
+      }
     }
   }
+  ${FRAGMENT_ITEM}
 `
 
+// ?? NOTE:  THis is not functionality we currently use in web-client
+// ?? the current (3/2022) expectations is that re-adding move the item
+// ?? to the top of the users list.  This is achieved with the upsert mutation
 export function itemUnArchive(itemId) {
   return requestGQL({
     query: itemUnArchiveQuery,