+++ src/common/api/queries/search-saved-items.js
@@ -3,9 +3,14 @@ import { gql } from 'graphql-request'
 import { FRAGMENT_ITEM } from 'common/api/fragments/fragment.item'
 
 const searchSavedItemsQuery = gql`
-  query SearchSavedItems($filter: SearchFilterInput, $pagination: PaginationInput, $term: String!) {
+  query SearchSavedItems(
+    $filter: SearchFilterInput
+    $pagination: PaginationInput
+    $sort: SearchSortInput
+    $term: String!
+  ) {
     user {
-      searchSavedItems(filter: $filter, pagination: $pagination, term: $term) {
+      searchSavedItems(filter: $filter, sort: $sort, pagination: $pagination, term: $term) {
         edges {
           cursor
           node {