+++ src/connectors/snowplow/actions/reader.js
@@ -151,5 +151,20 @@ export const readerActions = {
       uiType: 'button'
     },
     expects: ['id', 'url']
+  },
+  'reader.display': {
+    eventType: 'engagement',
+    batchEntityTypes: ['ui'],
+    eventData: {
+      uiType: 'button'
+    },
+    expects: ['value', 'label']
+  },
+  'reader.goback': {
+    eventType: 'engagement',
+    entityTypes: ['ui'],
+    eventData: {
+      uiType: 'button'
+    }
   }
 }