+++ src/containers/read/read.state.js
@@ -3,6 +3,7 @@ import { v4 as uuid } from 'uuid'
 import { arrayToObject } from 'common/utilities'
 import { localStore } from 'common/utilities/browser-storage/browser-storage'
 
+// V3 actions
 import { ARTICLE_ITEM_REQUEST } from 'actions'
 import { ARTICLE_ITEM_SUCCESS } from 'actions'
 import { ARTICLE_ITEM_FAILURE } from 'actions'
@@ -18,6 +19,7 @@ import { ANNOTATION_DELETE_REQUEST } from 'actions'
 import { ANNOTATION_DELETE_SUCCESS } from 'actions'
 import { ANNOTATION_DELETE_FAILURE } from 'actions'
 
+// Settings actions
 import { HYDRATE_DISPLAY_SETTINGS } from 'actions'
 import { UPDATE_LINE_HEIGHT } from 'actions'
 import { UPDATE_COLUMN_WIDTH } from 'actions'
@@ -47,11 +49,35 @@ import { deriveListItem } from 'common/api/derivers/item'
 
 import { HYDRATE } from 'actions'
 
+import { SNOWPLOW_SEND_EVENT } from 'actions'
+
+// Client API actions
+import { READ_ITEM_REQUEST } from 'actions'
+import { READ_ITEM_SUCCESS } from 'actions'
+import { READ_ITEM_FAILURE } from 'actions'
+
+import { HIGHLIGHT_SAVE_REQUEST } from 'actions'
+import { HIGHLIGHT_SAVE_SUCCESS } from 'actions'
+import { HIGHLIGHT_SAVE_FAILURE } from 'actions'
+
+import { HIGHLIGHT_DELETE_REQUEST } from 'actions'
+import { HIGHLIGHT_DELETE_SUCCESS } from 'actions'
+import { HIGHLIGHT_DELETE_FAILURE } from 'actions'
+
+import { createHighlight } from 'common/api'
+import { deleteHighlight } from 'common/api'
+
 /** ACTIONS
  --------------------------------------------------------------- */
+// v3 actions
 export const itemDataRequest = (itemId) => ({ type: ARTICLE_ITEM_REQUEST, itemId }) //prettier-ignore
 export const saveAnnotation = ({ itemId, quote, patch }) => ({ type: ANNOTATION_SAVE_REQUEST, item_id: itemId, quote, patch }) //prettier-ignore
 export const deleteAnnotation = ({ itemId, annotation_id }) => ({ type: ANNOTATION_DELETE_REQUEST, item_id: itemId, annotation_id }) //prettier-ignore
+// client-api actions
+export const readItemRequest = (itemId) => ({ type: READ_ITEM_REQUEST, itemId }) //prettier-ignore
+export const saveHighlightRequest = ({ itemId, quote, patch }) => ({ type: HIGHLIGHT_SAVE_REQUEST, itemId, quote, patch }) //prettier-ignore
+export const deleteHighlightRequest = ({ id }) => ({ type: HIGHLIGHT_DELETE_REQUEST, id }) //prettier-ignore
+// Settings actions
 export const updateLineHeight = (lineHeight) => ({ type: UPDATE_LINE_HEIGHT, lineHeight }) //prettier-ignore
 export const updateColumnWidth = (columnWidth) => ({ type: UPDATE_COLUMN_WIDTH, columnWidth }) //prettier-ignore
 export const updateFontSize = (fontSize) => ({ type: UPDATE_FONT_SIZE, fontSize }) //prettier-ignore
@@ -175,7 +201,9 @@ export const readSagas = [
   takeEvery(UPDATE_LINE_HEIGHT, saveDisplaySettings),
   takeEvery(UPDATE_COLUMN_WIDTH, saveDisplaySettings),
   takeEvery(UPDATE_FONT_SIZE, saveDisplaySettings),
-  takeEvery(UPDATE_FONT_TYPE, saveDisplaySettings)
+  takeEvery(UPDATE_FONT_TYPE, saveDisplaySettings),
+  takeEvery(HIGHLIGHT_SAVE_REQUEST, highlightSaveRequest),
+  takeEvery(HIGHLIGHT_DELETE_REQUEST, highlightDeleteRequest)
 ]
 
 /* SAGAS :: SELECTORS
@@ -272,6 +300,39 @@ function* annotationDeleteRequest({ item_id, annotation_id }) {
   }
 }
 
+function* highlightSaveRequest({ itemId, quote, patch }) {
+  try {
+    const highlight = {
+      version: 2,
+      itemId,
+      quote,
+      patch
+    }
+
+    const storedHighlights = yield select(getAnnotations)
+    const annotations = [...storedHighlights, highlight]
+
+    const data = yield call(createHighlight, highlight)
+
+    if (data) return yield put({ type: HIGHLIGHT_SAVE_SUCCESS, annotations })
+  } catch (error) {
+    yield put({ type: HIGHLIGHT_SAVE_FAILURE, error })
+  }
+}
+
+function* highlightDeleteRequest({ id }) {
+  try {
+    const storedAnnotations = yield select(getAnnotations)
+    const annotations = storedAnnotations.filter(i => i.id !== id) //prettier-ignore
+
+    const data = yield call(deleteHighlight, id)
+
+    if (data) return yield put({ type: HIGHLIGHT_DELETE_SUCCESS, annotations })
+  } catch (error) {
+    yield put({ type: HIGHLIGHT_DELETE_FAILURE, error })
+  }
+}
+
 function redirectToList() {
   if (document.location.href.indexOf('/read/') !== -1) {
     if (global.history.length > 1) {
@@ -297,6 +358,13 @@ function* saveDisplaySettings({ type, ...settings }) {
   yield Object.keys(settings).forEach((val) => {
     localStore.setItem(val.toString(), settings[val])
   })
+
+  const identifier = 'reader.display'
+  const data = Object.keys(settings).map((label) => ({
+    value: settings[label]?.toString(),
+    label
+  }))
+  yield put({ type: SNOWPLOW_SEND_EVENT, identifier, data })
 }
 
 /** ASYNC REQUESTS